+++ test/foundry/invariants/InvariantTestBase.sol
@@ -204,10 +204,10 @@ contract InvariantTestBase is LancaTest {
         s_parentPool.setMinDepositAmount(MIN_DEPOSIT_AMOUNT);
         s_parentPool.setMinWithdrawalAmount(MIN_WITHDRAWAL_AMOUNT);
         s_parentPool.grantRole(s_parentPool.LANCA_KEEPER(), s_lancaKeeper);
-        s_parentPool.setLurScoreSensitivity(uint64(5 * USDC_TOKEN_DECIMALS_SCALE));
+        s_parentPool.setLurScoreSensitivity(uint128(5 * USDC_TOKEN_DECIMALS_SCALE));
         s_parentPool.setScoresWeights(
-            uint64((7 * USDC_TOKEN_DECIMALS_SCALE) / 10),
-            uint64((3 * USDC_TOKEN_DECIMALS_SCALE) / 10)
+            uint128((7 * USDC_TOKEN_DECIMALS_SCALE) / 10),
+            uint128((3 * USDC_TOKEN_DECIMALS_SCALE) / 10)
         );
         s_parentPool.setLiquidityCap(LIQUIDITY_CAP);
         s_parentPool.setAverageConceroMessageFee(AVERAGE_CONCERO_MESSAGE_FEE);
